requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
dj-database-url==0.5.0
Django==4.1
django-channels==0.7.0
django-cors-headers==3.13.0
django-developer-panel==0.1.2
django-filter==22.1
django-livesync==0.5
django-on-heroku==1.1.2
djangorestframework==3.13.1
h11==0.13.0
h2==4.1.0
hiredis==2.0.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
incremental==21.3.0
jsonfield==3.1.0
Markdown==3.3.7
msgpack==1.0.4
oauthlib==3.2.0
priority==2.0.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
watchdog==2.1.9
websocket-client==1.3.3
whitenoise==6.2.0
wsproto==1.1.0
zipp==3.8.0
zope.interface==5.4.0
